initialization: Add security
- Configure DynamoDB as the storage for user-related data
- Configure Spring Security JWT authentication and authorization, with CSRF and custom error handling
- Secure the endpoint to retrieve information about a board game from boardgamegeek.com

NOTES

Up to this point, I have been trying to implement a hybrid approach to transparent error handling, declaring checked exceptions in methods that may not always return the expected result, and  using the utility methods provided by Atlassian Fugue to cover for the shortcomings of such approach (especially for parameters that expect lambda functions, which generally aren't expected to throw checked exceptions).
The idea was to use the functionality natively available in Java to enforce honest signatures for all methods, forcing the consumers to handle the possible anomalous conditions.
However, even though up till now such approach has proven to work,its shortcomings have become evident. Chief among these:

- It's necessary to declare many global exceptions, as each use case may define different conditions that impede further processing. This increase the types and verbosity in the code.
- The functional error handling utilities (Either, Try, etc.) have to be used anyway, in order to reduce the verbosity of the code, especially the try-catch blocks.
- The checked exceptions buble up, forcing successive wrappings, which further contributes to verbosity and doesn't add anything to code clarity.

For these reasons, I'm abandoning this moderate, hybrid approach. The next step is to refactor the existing code to adopt the functional approach, and continue with that from then and on.